remove function to devToolUtils
